Added undo/redo buttons. Fixed bug with Piece.mesh
Spamming redo when the move sequence included a pawn promotion was causing bugs because the queen's .mesh property was not being set until the animation was complete. The fix was to create the mesh and set the queen's .mesh property immediately in the promotion logic, and only make it appear (add it to piecesContainer) when the animation is finished.